+++ app/schemas/v2/donation.py
@@ -1,0 +1,5 @@
+from app.database.models import DonationBase, Player
+
+
+class DonationCascade(DonationBase):
+    player: Player
