+++ app/schemas/v2/player.py
@@ -1,0 +1,11 @@
+from app.database.models import Donation, PlayerBase, Whitelist, WhitelistBan
+
+
+class PlayerCascade(PlayerBase):
+    whitelists: list[Whitelist]
+    whitelists_issued: list[Whitelist]
+
+    whitelist_bans: list[WhitelistBan]
+    whitelist_bans_issued: list[WhitelistBan]
+
+    donations: list[Donation]
