+++ app/schemas/v2/whitelist.py
@@ -1,0 +1,11 @@
+from app.database.models import Player, WhitelistBanBase, WhitelistBase
+
+
+class WhitelistCascade(WhitelistBase):
+    player: Player
+    admin: Player
+
+
+class WhitelistBanCascade(WhitelistBanBase):
+    player: Player
+    admin: Player
